fix: refactor out admin feedback logic to own container
